% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdg_norm.R
\name{bdg_norm}
\alias{bdg_norm}
\title{bdg_norm (Reference sample based normalization)}
\usage{
bdg_norm(
  bridge.ls,
  data.ls,
  between.plate.method = "median",
  from_assay = "npx",
  save_assay = "inlot_normed"
)
}
\arguments{
\item{bridge.ls}{list of bridge.summarizedexperiment objs using pull_bdg()}

\item{data.ls}{list of summarizedexperiment objs using read_npx().
normalized assay slot will be added to each of the object in the list.}

\item{between.plate.method}{method to set the inter plate reference value of bridging samples
using max can garantee no negative value of the normalized data}

\item{from_assay}{select assay slot to be normalized}

\item{save_assay}{name the assay slot for normalized data}
}
\description{
normalize by adjust samples' value per plate to make that
the reference sample among all plates are the same. option to
pick using median, max, mean
}
