% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmb_npx_se.R
\name{cmb_npx_se}
\alias{cmb_npx_se}
\title{cmb_npx_se (Merge a list of npx summarizedexperiment to one)}
\usage{
cmb_npx_se(se_list)
}
\arguments{
\item{se_list}{list of npx summarizedexperiment obj
rowData will be removed, common colData}
}
\value{
combined summarizedexperiment obj
}
\description{
combine multiple npx summarizedexperiment obj
only common analyt would reimain in the merged object,
the LOD stored in the merged object will be the mean from all
individual npx objects
}
