% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_bdg.R
\name{pull_bdg}
\alias{pull_bdg}
\title{pull_bdg (creat summarizedexperiment of only reference sampls from a list of summarizedexperiments by identifing patterns)}
\usage{
pull_bdg(f_list, pattern = "hd", fields = "Assay")
}
\arguments{
\item{f_list}{list of summarizeexperiment objects}

\item{pattern}{sting pattern used for regex to identify samples}

\item{fields}{colData name from which to identify sample}
}
\value{
list of summarizedexperiment obj
}
\description{
pull bridging sample from list of summarizeexperiment object by pattern string
get bridging sample only before normalization by plate
}
