% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{read_npx}
\alias{read_npx}
\title{read_npx (read onlink npx file to summarizedexperiment)}
\usage{
read_npx(f, lot = "default", startrow = 8)
}
\arguments{
\item{f}{a string file path}

\item{startrow}{8 for npx, 9 for quant}

\item{lot:}{optional parameter to record lot information}
}
\value{
a summarizedexperimet object.
rowData is the feature and ctrl deviation,
colData is the unique_id and subject id given by the operator
}
\description{
read npx raw file to create a summarizedexperiment
uniportID and LOD in rowfeatures
file name, ctrl and qc warning are in the colData
}
