% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_npx_csv.R
\name{save_npx_csv}
\alias{save_npx_csv}
\title{save_npx_csv (same summarizedexpermient to csv)}
\usage{
save_npx_csv(se, pre_fix, assay2save = c("npx_default", "npx_inlot_normed"))
}
\arguments{
\item{se}{npx summarizedexpermient object}

\item{pre_fix}{string of the file path}

\item{assay2save}{vector string of which assay slot to save}
}
\description{
save summarizedexpermient to csv files
}
